fix(api): serialize a group's permissions as a number, not a flags string
Found by running the upgraded API against a real pre-groups database. The API
applies JsonStringEnumConverter globally, so GroupDto.Permissions went over the
wire as "ManageRooms, ManageUsers, ManagePlatform". The web does bit arithmetic
on that field (permissions & bit, permissions ^ bit), which silently yields 0 on
a string: every checkbox in the Groups tab would render unchecked, and toggling
one would send NaN.

GroupDto.Permissions is now an int. GroupInput already accepted a number, so the
request shape is unchanged.

Controller tests assert on the returned object and never serialize it, which is
exactly how this escaped 1,124 unit tests. GroupDtoSerializationTests serializes
through the app's own JsonConfig.

+++ tests/Diariz.Api.Tests/GroupsControllerTests.cs
@@ -41,7 +41,7 @@ public class GroupsControllerTests
         await sut.Create(new GroupInput("Engineering", "The eng team", "users", "#5C6BC0", PlatformPermission.ManageRooms));
 
         var groups = await sut.List();
-        Assert.Contains(groups, g => g.Name == "Engineering" && g.Permissions == PlatformPermission.ManageRooms);
+        Assert.Contains(groups, g => g.Name == "Engineering" && g.Permissions == (int)PlatformPermission.ManageRooms);
     }
 
     [Fact]
